rename more things to snake_case
